libs/clist: Move into internal
